add jobs and mail

+++ app/Http/Controllers/ArticleController.php
@@ -9,10 +9,14 @@ use Illuminate\Http\Request;
 class ArticleController extends Controller
 {
     public function create(){
+        $this->authorize('create', [self::class]);
         return view('articles/addArticle');
     }
 
     public function store(Request $request){
+
+        $this->authorize('create', [self::class]);
+
         $request->validate([
             'name' => 'required',
             'annotation' => 'required|min:10'
@@ -28,12 +32,13 @@ class ArticleController extends Controller
 
     public function view($articleId){
         $article = Article::findOrFail($articleId);
-        $comments = Comment::where('article_id', $articleId)->latest()->get();
+        $comments = Comment::where([['article_id', $articleId],['accept', 1]])->latest()->paginate(5);
         return view('articles/viewArticle', ['article' => $article, 'comments' => $comments]);
     }
 
     public function edit($id){
         $article = Article::FindOrFail($id);
+        $this->authorize('update', [$article, self::class]);
         return view('articles/editArticle', ['article'=>$article]);
     }
 
@@ -43,6 +48,7 @@ class ArticleController extends Controller
             'annotation' => 'required|min:10'
         ]);
         $article = Article::FindOrFail($id);
+        $this->authorize('update', [$article, self::class]);
         $article->name = request('name');
         $article->date = request('date');
         $article->shortDesc = request('annotation');
@@ -53,6 +59,7 @@ class ArticleController extends Controller
 
     public function destroy($id){
         $article = Article::FindOrFail($id);
+        $this->authorize('delete', [$article, self::class]);
         Comment::where('article_id', $id)->delete();
         $article->delete();
         return redirect('/');
